# black-on-white/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Черным по белому",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/article": {
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Создание новости",
                "operationId": "cb0b49ea078206c41ade76590662aa34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "text": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "article_type_id": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Аутентификация пользователя",
                "description": "Endpoint для аутентификации пользователя.",
                "operationId": "4ed2d4ad482dbe329dc4b3dd1e97ba7c",
                "requestBody": {
                    "description": "Учетные данные пользователя для аутентификации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "email@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized: неверный email или пароль",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Успешная аутентификация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "email"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": null
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "created_at"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "updated_at"
                                                },
                                                "is_admin": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "api_token": {
                                                    "type": "string",
                                                    "example": "api_token"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "authorisation": {
                                            "properties": {
                                                "token": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}